storage: pass storagepb.Liveness structs by value
This commit switches use of `storagepb.Liveness` from pointers
to values. Passing the struct by reference everywhere was the
cause of a number of allocations.

Most importantly, it was causing `NodeLiveness.getLivenessLocked`
to allocate, which was called in a number of mutex critical sections
including in `Replica.leaseStatus`. `Replica.leaseStatus` is called
by `Replica.leaseGoodToGo`, which we often see as a source of mutex
contention in read-heavy workloads.

Release note: None