rowexec: use a TableReader's ctx for the underlying Fetcher
Before this patch, the fetcher under a table reader was using a ctx
different than the TableReader's - a ctx without the TR's log tag. I
apparently thought that's a good idea back in the day for some reason.
But now I think there's no reason for it. Also, it's in my way. So I'm
making the fetcher simply use the TR's ctx.

Release note: None