storage: Move ExternalStorage creation to factory-like pattern
We'd like to easily change the interface for `MakeExternalStorage`
and `ExternalStorageFromURI` since they are the two ways we create
`ExternalStorage`, and we'd like to add a service that it those
objects can use.

This PR includes the plumbing to ensure that adding a file sharing
service to `ExternalStorage` is easy to do.

Release note: None