sql: use LeafTxn for some gateway flows
Before this patch, a DistSQL flow running on its gateway node would use
the RootTxn for all its processors for row-based flows / all of its
operators for vectorized flows if there are no remote flows. Some of
these processors/operator can execute concurrently with one another.
RootTxns don't support concurrent requests (see #25329), resulting in
some reads possibly missing to see the transaction's own writes.

This patch fixes things by using a LeafTxn on the gateway in case
there's concurrency on the gateway or if there's any remote flows. In
other words, the Root is used only if there's no remote flows and no
concurrency. This is sufficient for supporting mutations (which need the
Root), because mutations force everything to be planned on the gateway
and so, thanks to the previous commit, there's no concurrency if that's
the case.

Fixes #40487
Touches #24798

Release justification: Fixes bad bugs.

Release note: Fix a bug possibly leading to transactions missing to see
their own previous writes (#40487).