sql: use RootTxn for all flows that don't have remotes
Before this patch, queries that we didn't intend to distributed were
resulting in a single gateway flow that was using the RootTxn. Queries
that we did intend to distribute but where we didn't end up actually
distributing (e.g. because all the data was local, in particular when
running on a single node) were using a LeafTxn.
This patch makes the behavior uniform - if we ended up with a single
flow, for whatever reason, use the RootTxn.

Release justification: Simplifies things for next commits.

Release note: None